workspaces: implcitely create and delete workspaces
Mimmicks i3. Note use with the SpaceId app to see active spaces in dock
Biggest difference to i3 is that macos inserts a new space (i+1) for
fullscreen apps, where i3 keeps it at the same position (i). This means
the index of all subsequent spaces is temporarily increased, AND the
origin is still accessible, unlike i3.